bit int改int

+++ common-lib/src/main/java/com/talentcard/common/pojo/ScenicEnjoyPO.java
@@ -17,11 +17,11 @@ public class ScenicEnjoyPO implements Serializable {
 
     private Long categoryId;
 
-    private Long educationId;
+    private Integer educationId;
 
-    private Long titleId;
+    private Integer titleId;
 
-    private Long quality;
+    private Integer quality;
 
     /**
      * 1：人才卡；2：人才类别；3：人才学历；4：职称；5：职业资格
